Move all CAS1 OASys ITs into Cas1OasysTest
I should have done this when moving the corresponding endpoints
